Manually remove `CHANGELOG` entry
This commit manually removes an entry from `CHANGELOG.next.toml`. We need
to do this because the head commit of `smithy-rs-release-0.100.x` did not
go through the GitHub release workflow since it was only a codegen-only
fix, and the workflow is the one that moves `[[aws-sdk-rust]]` entries
from `CHANGELOG.next.toml` and `SDK_CHANGELOG.next.json`.