Fix broken link in README.md (#3227)
## Motivation and Context
The link to the design documentation in the root-level README is broken.
It produces a 404 error.

## Description
I changed the link in the README from
`https://awslabs.github.io/smithy-rs/design` to
`https://smithy-lang.github.io/smithy-rs/design/`

## Testing
I clicked the new link and it worked :)

## Checklist
N/A

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._