Fix stalled stream detection
A number of small issues in StalledStreamDetection came together to make it ineffective in practice:
1. We didn't throughput `0` on poll pending. This means that if the stream goes to poll pending, the calculate throughput can never get to 0.
2. Because we always considered the _entire_ throughput log, an actually stalled stream creates a pathological case where most of the log is full of fast moving data, but we are very slow to evict it and detect slowness. With a 426 length log, it would effecively take 426 seconds for it to actually get to 0.

For these two issues, I introduced a check-window (default 1 second) in the throughput log. We only consider this much data when making the calculation. To avoid undeseriable interactions with the check interval, I reduced the check interval to 500ms. I think this is likely to be better in practice at detecting stalled streams.

As a coincidence, this makes all the math exact so the tests are now precise without floating epsilons.

Finally, in an unrelated issue, the `check_interval` wasn't actually being used in the call to sleep, I fixed that as well and added a test.

In doing this, I tried to also simplify sine wave test into a test that could be logically reasoned about. The previous test was producing nonsense data (integer input to `sine` which is in radians), e.g.
I replaced it with several tests that send data at one rate, then suddenly stop sending data, simulating a stalled stream. These have the nice property that it's easy to determine mathematically when the stream should stall.