Serialize zero values
Removes logic from RustWriter that was used to not serialize
zero values, ie. 0 for numbers and false for booleans, if
that was also the default value for the member. With this change,
the only time the default value for a member is not serialized is
if the member is also `@clientOptional`.

The Blob type was also updated to implement std::Default, since
SymbolVisitor treats `""` as a RustDefault.

It _seems_ like this logic was originally added to fix issues with
[S3](https://github.com/smithy-lang/smithy-rs/issues/102) which
are technically modeling issues and have since been fixed.
However, this change impacts all the protocols and will cause
many members to now be sent over the wire that previously weren't.