Remove some defaults for s3 control
Adds the members of PublicAccessBlockConfiguration to
the RemoveDefaults customization. These members are
not meant to have a default, as you are meant to be able
to configure one member at a time - if you're just trying
to turn BlockPublicAcls on, you don't want to accidentally
turn off IgnorePublicAcls. This change makes these members
nullable and is a break-fix.

Also refactors the map of shapes to remove defaults from
to avoid having to call .shapeId everywhere.