Remove some defaults for s3 control (#3246)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->

S3 Control's PublicAccessBlockConfiguration is incorrectly modeled, and
all the members should be nullable. These members are not meant to have
a default, as you are meant to be able to configure one member at a time
- if you're just trying to turn BlockPublicAcls on, you don't want to
accidentally turn off IgnorePublicAcls. This change makes these members
nullable and is a break-fix.

## Description
<!--- Describe your changes in detail -->

Adds the members of PublicAccessBlockConfiguration to the RemoveDefaults
customization.
Also refactors the map of shapes to remove defaults from to avoid having
to call .shapeId everywhere.

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
- [x] Generated the client and looked at it

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] I have updated `CHANGELOG.next.toml` if I made changes to the AWS
SDK, generated SDK code, or SDK runtime crates

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._