Ignore actual default value instead of 0/false
Previously, there was logic used to ignore the default 0/false values
for numbers/booleans when serializing members if they weren't boxed.
This was to fix issues that occured when upstream models didn't
properly box shapes that were meant to be optional, and for the
most part worked because services would just fill in the default if
it wasn't passed. However, with Smithy 2.0, models may have defaults
!= 0/false, but the codegenerator still ignores the zero value.

This commit makes it so that the actual default value for the member
is ignored for booleans and numbers, instead of just 0/false. It
also updates serialization for http bindings so that headers and
query parameters with 0/false values aren't ignored if they are
optional parameters.