Move `aws-http` types into `aws-runtime` (#3355)
This issue addresses a semver compatibility problem similar to the one
described in https://github.com/smithy-lang/smithy-rs/pull/3318, except
for the storable types in the aws-http crate. I opted to move all of
aws-http into aws-runtime since there was so little in there.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._