Update WASI HTTP client
Added a Builder to allow for future configuration to be baked in in a
backwards compatible manner. Also updated some of the error handling.

Updated the associated tests to use the new Builder and to not send a
network call. Sending the actual network call will be moved to the
canary in another commit.

Undoing some autoformatting in the Dockerfile

Lint fixes