Allow S3 Express to use CRC32 for default request checksum (#3432)
## Motivation and Context
Changes the default request checksum to CRC32 for S3 Express.

## Description
This PR modifies the default request checksum to CRC32 for S3 Express.
Note that if a customer specifies a custom checksum algorithm via such
as
[checksum_algorithm](https://docs.rs/aws-sdk-s3/latest/aws_sdk_s3/operation/put_object/struct.PutObjectInput.html#structfield.checksum_algorithm),
then the PR does not affect anything and respects the customer-specified
checksum algorithm.

Otherwise, if an operation is annotated with a `requestChecksumRequired`
trait and set to true, the default checksum becomes relevant. The
default algorithm is currently `MD5` and S3 Express requires it to be
`CRC32`. The code changes in this PR reflect that requirement.

Implementation details include
- `RequestChecksumInterceptor` is a single-sourced place to handle
request checksum, but since it is a common place for all services we
cannot add S3-Express-specific logic in there to configure the default
checksum. Instead, we have introduced a slight more abstract concept of
overriding the default checksum via `DefaultRequestChecksumOverride`.
`RequestChecksumInterceptor` grabs it from a `ConfigBag` and uses it (if
found) to configure the default checksum.
- We deferred a point of adding a checksum to a request body to
`RequestChecksumInterceptor::modify_before_signing` from
`RequestChecksumInterceptor::modify_before_retry_loop`. This is because
we won't know whether S3 Express checksum default actually kicks until
an endpoint is fully resolved. `modify_before_retry_loop` occurs before
`client::orchestrator::endpoints::orchestrate_endpoint` so we need to
choose a different interceptor method that runs after it.
- I abandoned having a separate S3-specific `RequestChecksumInterceptor`
as it was more complicated than this approach. The wrinkle is that if we
have an S3-specific `RequestChecksumInterceptor`, then it needs to
either a) disable [the general
RequestChecksumInterceptor](https://github.com/smithy-lang/smithy-rs/blob/f7be5f801504e7edf5cb6b731f3400c37b298abc/aws/rust-runtime/aws-inlineable/src/http_request_checksum.rs#L79-L100)
entirely or b) only change the checksum default and delegate the rest to
the general `RequestChecksumInterceptor`. In both cases the general
`RequestChecksumInterceptor` must be disabled in runtime components,
otherwise a request checksum related header will incorrectly be added
twice. To disable the general `RequestChecksumInterceptor` using
[disable_interceptor](https://github.com/smithy-lang/smithy-rs/blob/main/rust-runtime/aws-smithy-runtime-api/src/client/interceptors.rs#L839),
we need to remove a generic parameter `AP` from
`RequestChecksumInterceptor` since we cannot enumerate all possible `AP`
when specifying it in `disable_interceptor`.
 
## Testing
Added integration tests for S3 Express default checksum to
`integration-tests/s3/tests/express.rs`.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._